Add Traversable instance for the AST.
We use it to "intern" strings in the AST so every name is stored exactly
once. This saves around 35% on a typical file. E.g. parsing all
toxcore/*.c files together takes up 9,961,472 bytes with interning, and
would take 15,335,424 bytes without interning. It takes 20,971,520 bytes
with String instead of Text (with sharing - without sharing it takes
21,790,720 bytes).

Also, for comparison: check-cimple takes around 0.25 seconds to parse all
those files to `Node String`. Then it takes about 5ms (0.005s) to intern
all the strings. We then store the entire AST in a contiguous block of
memory, which would take 150ms if we didn't preserve sharing (so the
interning we did would be useless), but actually takes about 1 second
(1000ms) to produce a contiguous data structure of around 10MiB.

Time spent in parsing and preparing the AST for analysis is thus around
1.2 seconds. We may later decide that this is not worth it since we're
not a long-running application and would likely never benefit from the
reduced GC pressure, and cache locality is unlikely to be much better
than with a regular GC'd data structure. It's an interesting experiment,
though :).