feat: Add check for locals that can be const.
Arrays are ignored because they can implicitly convert to pointer types
and be written to that way (we have no way of knowing this without a
type check).

`va_list` is ignored because it's an array.

Local pointers are ignored for now because I'm not sure we should make
them all const (as opposed to making the pointee const, which is
definitely what we should do, but this analysis doesn't check for that).