feat: More improvements and simplifications to type check.
The type check is now in theory commutative, so

`unify a b` == `unify b a`.

Whether this is really true, I don't know yet. I'll do a formal proof at
some point later. For now, this works well. One problem is that the
pattern matching order matters, which seems to indicate it might not
actually be commutative.