test: Add a test to make sure arrays are treated as expected.
This was done correctly, but we hadn't documented why arrays are
ignored. Ignoring them causes false negatives, but the false positives
are somewhat difficult to avoid. See the tests for two cases in which we
don't want to reduce the scope of arrays but we would want to reduce the
scope of pointers (which are indistinguishable at the use-site).