feat: Check that enum `from_int` functions are correct.
This is the first "semantic formatter" style linter. It's pretty simple
and just checks that all enumerators are present and there is a sane
default in the `${enum}_from_int` functions. There is no compiler
warning we can use for this, because the int can contain anything.
Currently, if there is an enumerator ending in `_INVALID`, then that is
the default, otherwise the first enumerator is the default for invalid
int values.